Rename register_tokenizer.rs to submit_tokenizer_job.rs
